Move widget inside module.
