+++ assets/js/modules/analytics-4/components/widgets/index.js
@@ -26,3 +26,4 @@ export { default as TopCitiesWidget } from './TopCitiesWidget';
 export { default as TopCountriesWidget } from './TopCountriesWidget';
 export { default as TopTrafficSourceWidget } from './TopTrafficSourceWidget';
 export { default as TopConvertingTrafficSourceWidget } from './TopConvertingTrafficSourceWidget';
+export { default as ConnectGA4CTAWidget } from './ConnectGA4CTAWidget';
